backend/requirements.txt
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
sentence-transformers
unstructured
pypdf
python-docx
flask
flask-cors
firebase-admin
supabase
faiss-cpu
unstructured[docx]
pytz
typing-extensions
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib